# src/main/java/ru/savimar/billing/openapi.yaml
openapi: 3.0.0
info:
  description: |
    This is a sample Petstore server.  You can find
    out more about Swagger at
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: Swagger Petstore
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/new445/notifications/1.0.0
  - url: 'https://petstore.swagger.io/v2'
...
paths:
...
  '/bill/{billId}':
    get:
      tags:
        - bill
      summary: Find bill by ID
      description: Returns a single bill
      operationId: getBillById
      parameters:
        - name: billId
          in: path
          description: ID of bill to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bill'
            application/xml:
              schema:
                $ref: '#/components/schemas/Bill'
        '400':
          description: Invalid ID supplied
        '404':
          description: bill not found

    post:
      tags:
        - bill
      summary: Updates a bill in the store with form data
      operationId: updatePetWithForm
      parameters:
        - name: billId
          in: path
          description: ID of bill that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '405':
          description: Invalid input

      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  description: Updated name of the bill
                  type: string
                status:
                  description: Updated status of the bill
                  type: string
    delete:
      tags:
        - bill
      summary: Deletes a bill
      operationId: deletebill
      parameters:

        - name: billId
          in: path
          description: bill id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid ID supplied
        '404':
          description: bill not found
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64

        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        bill:
          $ref: '#/components/schemas/User'
      xml:
        name: User
...
    Bill:
      type: object


      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'

      xml:
        name: Bill
...